+++ Changelog.md
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+Version (NEXT)
+--------------
+* fixed relative link functions in markdown api for empty namespace elements
+
 Version 0.32.0
 --------------
 * added model navigation functions for state machines